{{ cookiecutter.target }}/backend/test/conftest.py
from {{ cookiecutter.project }}.testing import FUNCTIONAL_TESTING
from collective.elastic.plone.testing import INTEGRATION_TESTING
from pytest_plone import fixtures_factory


pytest_plugins = ["pytest_plone"]


globals().update(
    fixtures_factory(
        (
            (FUNCTIONAL_TESTING, "functional"),
            (INTEGRATION_TESTING, "integration"),
        )
    )
)