{{ cookiecutter.target }}/backend/pyproject.toml
[project]
name = "{{ cookiecutter.project }}"
version = "1.0.0dev0"
description = "Plone Distribution {{ cookiecutter.project}}"
readme = "README.md"
maintainers = [
    {name = "John Doe", email = "jd@example.com"},
]
authors = [
    {name = "Max Msutermann"},
]
requires-python = ">=3.11"
license = { text = "GPL 2.0" }
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Plone",
    "Framework :: Plone :: Addon",
    "Framework :: Plone :: 6.1",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Development Status :: 5 - Production/Stable",
]
dependencies = [
    "plone.distribution",
    "plone.volto",
]

[project.urls]
PyPI = "https://pypi.python.org/pypi/{{ cookiecutter.project}}"
Changelog = "https://github.com/collective/{{ cookiecutter.project}}/blob/main/CHANGES.rst"
Source = "https://github.com/collective/{{ cookiecutter.project}}"
Issues = "https://github.com/collective/{{ cookiecutter.project}}/issues"

[project.optional-dependencies]
development = [
    "plone.app.contenttypes[test]",
    "pytest-plone",
    "pdbpp",
    "Products.PrintingMailhost",
    "waitress_fastlisten",
]
production = [
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]

[tool.isort]
profile = "plone"

[tool.codespell]
ignore-words-list = "discreet,"
skip = './venv/*'

[tool.check-manifest]
ignore = [
    ".editorconfig",
    ".pre-commit-config.yaml",
    "tox.ini",
    "mypy.ini",
    ".flake8",
    "mx.ini",

]
